fix connectivity notification making sounds on change
